Pull the completed Color Drift feature from wisp-restructure: wisp rolling
drift engine with lamp-animated long fades, app Color Drift panel (how-often
+ fade sliders, debounced setDrift), NVS persistence, and wispStatus echo.

Wire change: MSG_OVERRIDE_COLORS fadeDurationMs widened 2->4 LE bytes
(FIXED_SIZE 23->25, numColors -> [24]) for ~1hr drift fades. No protocol
version bump — the whole fleet (lamps + wisp) must be reflashed together.
RESTORE_COLORS / OVERRIDE_BRIGHTNESS keep uint16.

Conflict resolution: wisp-restructure branched pre-split, so it edited the
old monolithic lamp_protocol.hpp and the inline show_receiver Pending structs.
Re-derived the uint32 change into the split protocol/override.hpp and the
extracted pending_slots.hpp (PendingOverrideColors/PendingRestoreColors);
kept mesh_link.hpp at HEAD; esp-modbus strip auto-deduped.

Verified: firmware build SUCCESS, native 425/425 (incl. new protocol uint32
round-trip tests), wisp build SUCCESS + 61/61, app 443/443.

+++ software/lamp-os/src/behaviors/configurator.hpp
@@ -50,14 +50,14 @@ class ConfiguratorBehavior : public AnimatedBehavior {
   // start). When `fadeDurationMs == 0` (or `targetColors` is empty) this
   // is effectively an instant set.
   void beginFade(const std::vector<Color>& targetColors,
-                 uint16_t fadeDurationMs);
+                 uint32_t fadeDurationMs);
 
   // ColorOverride/BLE color writes need to read back the
   // fade-tracking state to decide whether the configurator is mid-fade.
   // Exposed publicly so the override modules can drive the transition
   // without owning their own per-pixel interpolation.
   uint32_t fadeStartMs() const { return fadeStartMs_; }
-  uint16_t fadeDurationMs() const { return fadeDurationMs_; }
+  uint32_t fadeDurationMs() const { return fadeDurationMs_; }
   bool fadeActive(uint32_t nowMs) const;
 
  private:
@@ -67,7 +67,7 @@ class ConfiguratorBehavior : public AnimatedBehavior {
   // pixel count changes (rare — only on configuration boot).
   std::vector<Color> fadeFromColors_;
   uint32_t fadeStartMs_ = 0;
-  uint16_t fadeDurationMs_ = 0;
+  uint32_t fadeDurationMs_ = 0;
 };
 
 }  // namespace lamp
